arroyo-worker: pull GRPC code out of engine.
